src/third_party/blink/web_tests/serial/serialPort_readable_cancel.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/unguessable_token.mojom.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/serial/serial.mojom.js"></script>
<script src="resources/serial-test-utils.js"></script>
<script>

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  // Select a buffer size smaller than the amount of data transferred.
  await port.open({ baudrate: 9600, buffersize: 64 });

  const reader = port.readable.getReader();
  const readPromise = reader.read();
  await reader.cancel();
  const { value, done } = await readPromise;
  assert_true(done);
  assert_equals(undefined, value);

  await port.close();
}, 'Can cancel while reading');

</script>
